Move `get_run_date` from `utils` to `render_report`
This function is only used in this module.
