+++ analysis/render_report.py
@@ -6,8 +6,10 @@ For more information about Jinja, see:
 import base64
 import collections
 import datetime
+import json
 import mimetypes
 
+import dateutil.parser
 from jinja2 import Environment, FileSystemLoader, StrictUndefined
 
 from analysis import utils
@@ -30,7 +32,7 @@ def main():
             # It's passed as a template variable so that we can format it consistently
             # with other template variables.
             "tpp_epoch_date": datetime.date(2009, 1, 1),
-            "run_date": utils.get_run_date(),
+            "run_date": get_run_date(),
             "from_date": {
                 "plot_from_2020": datetime.date(2020, 2, 1),
                 "plot_from_2016": datetime.date(2016, 1, 1),
@@ -65,6 +67,21 @@ ENVIRONMENT.filters["date_format"] = utils.date_format
 def render_report(data):
     template = ENVIRONMENT.get_template("report_template.html")
     return template.render(data)
+
+
+def get_log():
+    return [
+        json.loads(line)
+        for line in (utils.OUTPUT_DIR / "query" / "log.json").read_text().splitlines()
+    ]
+
+
+def get_run_date():
+    by_event = {d["event"]: d for d in get_log()}
+    timestamp = by_event.get("finish_executing_sql_query", {}).get(
+        "timestamp", "9999-01-01T00:00:00"
+    )
+    return dateutil.parser.parse(timestamp)
 
 
 def group_plots(*paths, suffix=".png"):
